Write game logic and added script to index.html
I finished writing the game logic for rock paper scissors, which takes
input from the user in a prompt and compares it to a random value given
to the computer between 1 and 3 (inclusive). The game goes on for 5
rounds, and after each round either the player or the computer ends up
with one point (or none if it's a tie). After all 5 rounds, whoever has
the highest score is declared the winner.