Add game logic to buttons with event listeners
I added onclick event listeners to each of the buttons. When clicked,
they ran a round of rock paper scissors, incrementing the score and
checking to see if either the computer or player had reached 5 points
yet. Each round, the scores would be displayed on the h2 below the boxes.
Once the game was over, the player would either be announced the winner
or the loser. This required refactoring the code and getting rid of the
while loop that was previously used to run the game. Instead, the game
relied on the player selecting the boxes to move the game forward.